# swagger.yaml
openapi: 3.0.3
info:
  title: Quiz Application API
  description: API specification for the AI-powered adaptive language learning quiz application. All `date-time` fields in this specification use RFC3339 timestamps and include timezone offsets (e.g., `2025-08-15T16:33:19Z` or `2025-08-15T12:33:19-04:00`). Servers MUST return timestamps with offsets and clients MUST parse them respecting the offset.
  version: 1.0.0
  contact:
    name: Quiz Team
paths:
...
  /v1/admin/worker/status:
    get:
      tags:
        - Admin
      summary: Get worker status
      description: Get the current status of the worker (admin only)
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Worker status retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkerStatus"
        "401":
          description: User not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: User not authorized (admin required)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/admin/worker/details:
    get:
      tags:
        - Admin
      summary: Get worker details
      description: Get detailed information about the worker (admin only)
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Worker details retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkerDetailsResponse"
        "401":
          description: User not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: User not authorized (admin required)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/admin/worker/pause:
    post:
      tags:
        - Admin
      summary: Pause worker
      description: Pause the worker (admin only)
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Worker paused successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          description: User not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: User not authorized (admin required)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/admin/worker/resume:
    post:
      tags:
        - Admin
      summary: Resume worker
      description: Resume the worker (admin only)
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Worker resumed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          description: User not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: User not authorized (admin required)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/admin/worker/trigger:
    post:
      tags:
        - Admin
      summary: Trigger worker run
      description: Trigger a manual worker run (admin only)
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Worker run triggered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          description: User not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: User not authorized (admin required)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/admin/worker/logs:
    get:
      tags:
        - Admin
      summary: Get worker logs
      description: Get worker activity logs (admin only)
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Worker logs retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkerLogsResponse"
        "401":
          description: User not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: User not authorized (admin required)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/admin/worker/ai-concurrency:
    get:
      tags:
        - Admin
      summary: Get worker AI concurrency statistics
      description: Get AI concurrency statistics from the worker (admin only)
      security:
        - cookieAuth: []
      responses:
        "200":
          description: AI concurrency stats retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkerAIConcurrencyStats"
        "401":
          description: User not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: User not authorized (admin required)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/admin/worker/users:
    get:
      tags:
        - Admin
      summary: Get worker user list with pause status
      description: Get list of users with their worker pause status (admin only)
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Worker users retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkerUserListResponse"
        "401":
          description: User not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: User not authorized (admin required)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/admin/worker/users/pause:
    post:
      tags:
        - Admin
      summary: Pause worker for a specific user
      description: Pause question generation for a specific user (admin only)
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserIdRequest"
      responses:
        "200":
          description: User paused successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: User not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: User not authorized (admin required)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/admin/worker/users/resume:
    post:
      tags:
        - Admin
      summary: Resume worker for a specific user
      description: Resume question generation for a specific user (admin only)
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserIdRequest"
      responses:
        "200":
          description: User resumed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: User not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: User not authorized (admin required)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/admin/worker/analytics/priority-scores:
    get:
      tags:
        - Admin
      summary: Get worker priority analytics
      description: Get priority analytics from the worker (admin only)
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Priority analytics retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PriorityAnalytics"
        "401":
          description: User not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: User not authorized (admin required)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/admin/worker/analytics/user-performance:
    get:
      tags:
        - Admin
      summary: Get worker user performance analytics
      description: Get user performance analytics from the worker (admin only)
      security:
        - cookieAuth: []
      responses:
        "200":
          description: User performance analytics retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserPerformanceAnalytics"
        "401":
          description: User not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: User not authorized (admin required)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/admin/worker/analytics/generation-intelligence:
    get:
      tags:
        - Admin
      summary: Get worker generation intelligence
      description: Get generation intelligence from the worker (admin only)
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Generation intelligence retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerationIntelligence"
        "401":
          description: User not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: User not authorized (admin required)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /v1/admin/worker/notifications/stats:
    get:
      tags:
        - Admin
      summary: Get notification statistics
      description: Get comprehensive notification statistics including sent count, error count, and user statistics (admin only)
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Notification statistics retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationStats"
        "401":
          description: User not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: User not authorized (admin required)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/admin/worker/notifications/errors:
    get:
      tags:
        - Admin
      summary: Get notification errors
      description: Get paginated list of notification errors with filtering options (admin only)
      security:
        - cookieAuth: []
      parameters:
        - name: page
          in: query
          description: Page number (1-based)
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: page_size
          in: query
          description: Number of errors per page
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: error_type
          in: query
          description: Filter by error type
          schema:
            type: string
            enum:
              [
                "smtp_error",
                "template_error",
                "user_not_found",
                "email_disabled",
                "other",
              ]
        - name: notification_type
          in: query
          description: Filter by notification type
          schema:
            type: string
            enum: ["daily_reminder", "test_email"]
        - name: resolved
          in: query
          description: Filter by resolution status
          schema:
            type: string
            enum: ["true", "false"]
      responses:
        "200":
          description: Notification errors retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkerNotificationErrorsResponse"
        "401":
          description: User not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: User not authorized (admin required)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/admin/worker/notifications/sent:
    get:
      tags:
        - Admin
      summary: Get sent notifications
      description: Get paginated list of sent notifications with filtering options (admin only)
      security:
        - cookieAuth: []
      parameters:
        - name: page
          in: query
          description: Page number (1-based)
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: page_size
          in: query
          description: Number of notifications per page
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: notification_type
          in: query
          description: Filter by notification type
          schema:
            type: string
            enum: ["daily_reminder", "test_email"]
        - name: status
          in: query
          description: Filter by status
          schema:
            type: string
            enum: ["sent", "failed", "bounced"]
        - name: sent_after
          in: query
          description: Filter notifications sent after this timestamp
          schema:
            type: string
            format: date-time-time
        - name: sent_before
          in: query
          description: Filter notifications sent before this timestamp
          schema:
            type: string
            format: date-time-time
      responses:
        "200":
          description: Sent notifications retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkerNotificationSentResponse"
        "401":
          description: User not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: User not authorized (admin required)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/admin/worker/notifications/force-send:
    post:
      tags:
        - Admin
      summary: Force send notification to user
      description: Force send a daily reminder notification to a specific user, bypassing normal time and date checks (admin only)
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
              properties:
                username:
                  type: string
                  description: Username of the user to send notification to
                  example: "testuser"
      responses:
        "200":
          description: Notification sent successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForceSendNotificationResponse"
        "400":
          description: Invalid request or user has notifications disabled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: User not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: User not authorized (admin required)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Email service not available
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /v1/admin/worker/daily/users/{userId}/questions/{date}:
    get:
      tags:
        - Worker Admin
      summary: Get daily questions for a specific user and date
      description: Retrieve all daily questions assigned to a specific user for a given date (admin only)
      security:
        - cookieAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          description: ID of the user
          schema:
            type: integer
        - name: date
          in: path
          required: true
          description: Date in YYYY-MM-DD format
          schema:
            type: string
            format: date
            example: "2024-01-15"
      responses:
        "200":
          description: Daily questions retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminDailyQuestionsResponse"
        "400":
          description: Invalid user ID or date format
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Failed to get daily questions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/admin/worker/daily/users/{userId}/questions/{date}/regenerate:
    post:
      tags:
        - Worker Admin
      summary: Regenerate daily questions for a specific user and date
      description: Clear and regenerate daily questions for a specific user and date (admin only)
      security:
        - cookieAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          description: ID of the user
          schema:
            type: integer
        - name: date
          in: path
          required: true
          description: Date in YYYY-MM-DD format
          schema:
            type: string
            format: date
            example: "2024-01-15"
      responses:
        "200":
          description: Daily questions regenerated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Invalid user ID or date format
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Failed to regenerate daily questions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
          description: Error code identifying the type of error
        message:
          type: string
          description: Human-readable error message
        error:
          type: string
          maxLength: 500
          description: Error message (for backward compatibility)
        details:
          type: string
          maxLength: 1000
          description: Additional error details
        cause:
          type: string
          description: Underlying error cause (included for error and fatal severity levels)
        retryable:
          type: boolean
          description: Whether the operation can be retried
        severity:
          type: string
          enum: [info, warn, error, fatal]
          description: Severity level of the error
...
    WorkerStatus:
      type: object
      properties:
        status:
          type: string
          enum: ["idle", "busy", "error"]
          description: Current status of the worker
        last_heartbeat:
          type: string
          format: date-time-time
          description: Timestamp of the last heartbeat from the worker
        error_message:
          type: string
          nullable: true
          description: Error message if the worker is in an error state
...
    MessageResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
...
    WorkerDetailsResponse:
      type: object
      additionalProperties: true
    WorkerLogsResponse:
      type: object
      required:
        - logs
      properties:
        logs:
          type: array
          items:
            type: object
            additionalProperties: true
...
    WorkerUserListResponse:
      type: object
      required:
        - users
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/WorkerUserStatus"
    WorkerAIConcurrencyStats:
      type: object
      properties:
        active_requests:
          type: integer
        max_concurrent:
          type: integer
        queued_requests:
          type: integer
        total_requests:
          type: integer
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: aiquiz_session
    sessionAuth:
      type: apiKey
      in: cookie
      name: aiquiz_session
      description: Session-based authentication using cookies
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API Key
      description: API key authentication for programmatic access. Use format "Bearer qapp_..."
    apiKeyQuery:
      type: apiKey
      in: query
      name: api_key
      description: API key authentication via URL query parameter for programmatic access. Use "?api_key=qapp_..."